Mv participant stats set to utils
